# services/api-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "osparc.io public API",
    "description": "osparc-simcore public API specifications",
    "version": "0.7.0"
  },
  "paths": {
...
    "/v0/studies": {
      "get": {
        "tags": [
          "studies"
        ],
        "summary": "List Studies",
        "description": "New in *version 0.5.0*",
        "operationId": "list_studies",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Study_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "studies"
        ],
        "summary": "Create Study",
        "description": "Create a study\n\nNew in *version 0.8.0*",
        "operationId": "create_study",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyNodes"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Study"
                }
              }
            }
          },
          "404": {
            "description": "Study not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorGet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v0/studies/{study_id}": {
      "get": {
        "tags": [
          "studies"
        ],
        "summary": "Get Study",
        "description": "New in *version 0.5.0*",
        "operationId": "get_study",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Study Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Study"
                }
              }
            }
          },
          "404": {
            "description": "Study not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorGet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v0/studies/{study_id}:clone": {
      "post": {
        "tags": [
          "studies"
        ],
        "summary": "Clone Study",
        "operationId": "clone_study",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Study Id"
            }
          },
          {
            "name": "x-simcore-parent-project-uuid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Simcore-Parent-Project-Uuid"
            }
          },
          {
            "name": "x-simcore-parent-node-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Simcore-Parent-Node-Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Study"
                }
              }
            }
          },
          "404": {
            "description": "Study not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorGet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v0/studies/{study_id}/ports": {
      "get": {
        "tags": [
          "studies"
        ],
        "summary": "List Study Ports",
        "description": "Lists metadata on ports of a given study\n\nNew in *version 0.5.0*",
        "operationId": "list_study_ports",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Study Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnePage_StudyPort_"
                }
              }
            }
          },
          "404": {
            "description": "Study not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorGet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorGet": {
        "properties": {
          "errors": {
            "items": {},
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "errors"
        ],
        "title": "ErrorGet",
        "example": {
          "errors": [
            "some error message",
            "another error message"
          ]
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Validation errors"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Links": {
        "properties": {
          "first": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First"
          },
          "last": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last"
          },
          "self": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Self"
          },
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next"
          },
          "prev": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prev"
          }
        },
        "type": "object",
        "required": [
          "first",
          "last",
          "self",
          "next",
          "prev"
        ],
        "title": "Links"
      },
...
      "OnePage_StudyPort_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StudyPort"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "OnePage[StudyPort]"
      },
...
      "Page_Study_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Study"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "offset": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Offset"
          },
          "links": {
            "$ref": "#/components/schemas/Links"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "links"
        ],
        "title": "Page[Study]"
      },
...
      "Study": {
        "properties": {
          "uid": {
            "type": "string",
            "format": "uuid",
            "title": "Uid"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "uid"
        ],
        "title": "Study"
      },
      "StudyNodes": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/_Node"
            },
            "type": "array",
            "minItems": 1,
            "title": "Nodes",
            "description": "Array of service keys and versions"
          }
        },
        "type": "object",
        "required": [
          "nodes"
        ],
        "title": "StudyNodes"
      },
      "StudyPort": {
        "properties": {
          "key": {
            "type": "string",
            "format": "uuid",
            "title": "Key name",
            "description": "port identifier name.Correponds to the UUID of the parameter/probe node in the study"
          },
          "kind": {
            "type": "string",
            "enum": [
              "input",
              "output"
            ],
            "title": "Kind"
          },
          "content_schema": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Schema",
            "description": "jsonschema for the port's value. SEE https://json-schema.org"
          }
        },
        "type": "object",
        "required": [
          "key",
          "kind"
        ],
        "title": "StudyPort",
        "example": {
          "content_schema": {
            "maximum": 5,
            "minimum": 0,
            "title": "Sleep interval",
            "type": "integer",
            "x_unit": "second"
          },
          "key": "f763658f-a89a-4a90-ace4-c44631290f12",
          "kind": "input"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "_Node": {
        "properties": {
          "service_key": {
            "type": "string",
            "pattern": "^simcore/services/((comp|dynamic|frontend))/([a-z0-9][a-z0-9_.-]*/)*([a-z0-9-_]+[a-z0-9])$",
            "title": "Service Key"
          },
          "service_version": {
            "type": "string",
            "pattern": "^(0|[1-9]\\d*)(\\.(0|[1-9]\\d*)){2}(-(0|[1-9]\\d*|\\d*[-a-zA-Z][-\\da-zA-Z]*)(\\.(0|[1-9]\\d*|\\d*[-a-zA-Z][-\\da-zA-Z]*))*)?(\\+[-\\da-zA-Z]+(\\.[-\\da-zA-Z-]+)*)?$",
            "title": "Service Version"
          }
        },
        "type": "object",
        "required": [
          "service_key",
          "service_version"
        ],
        "title": "_Node"
      }
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}